Move some database related classes out
